Add VisaoCategorias and VisaoTarefas classes

+++ ArquivoCategoria.java
@@ -1,4 +1,5 @@
 import aed3.*;
+import java.util.ArrayList;
 
 public class ArquivoCategoria extends aed3.Arquivo<Categoria> {
 
@@ -21,26 +22,31 @@ public class ArquivoCategoria extends aed3.Arquivo<Categoria> {
     }
 
     public Categoria read(String nome) throws Exception {
-        ParNomeId pci = indiceIndiretoNome.read(new ParNomeId(nome)).get(0);
-        if (pci == null) {
+        ArrayList<ParNomeId> pcis = indiceIndiretoNome.read(new ParNomeId(nome));
+        if (pcis.isEmpty()) {
             return null;
         }
-        return read(pci.getId());
+        return super.read(pcis.get(0).getId());
     }
     
     public boolean delete(String nome) throws Exception {
-        ParNomeId pci = indiceIndiretoNome.read(new ParNomeId(nome)).get(0);
-        if (pci != null) {
-            if (delete(pci.getId())) {
-                return indiceIndiretoNome.delete(new ParNomeId(nome, pci.getId()));
-            }
+        ArrayList<ParNomeId> pcis = indiceIndiretoNome.read(new ParNomeId(nome));
+        if (pcis.isEmpty()) {
+            return false;
+        }
+        ParNomeId pci = pcis.get(0);
+        if (super.delete(pci.getId())) {
+            return indiceIndiretoNome.delete(new ParNomeId(nome, pci.getId()));
         }
         return false;
     }
 
     @Override
     public boolean update(Categoria novaCategoria) throws Exception {
-        Categoria categoriaAntiga = read(novaCategoria.getNome());
+        Categoria categoriaAntiga = read(novaCategoria.getId());
+        if (categoriaAntiga == null) {
+            return false;
+        }
         if (super.update(novaCategoria)) {
             if (!novaCategoria.getNome().equals(categoriaAntiga.getNome())) {
                 indiceIndiretoNome.delete(new ParNomeId(categoriaAntiga.getNome(), categoriaAntiga.getId()));
@@ -49,5 +55,13 @@ public class ArquivoCategoria extends aed3.Arquivo<Categoria> {
             return true;
         }
         return false;
+    }
+
+    public ArrayList<Categoria> readAll() throws Exception {
+        ArrayList<Categoria> categorias = new ArrayList<>();
+        for (ParNomeId pci : indiceIndiretoNome.read(null)) {
+            categorias.add(super.read(pci.getId()));
+        }
+        return categorias;
     }
 }
